+++ tests/AzureTackleAnalyzer.Tests/AssemblyInfo.fs
@@ -4,20 +4,20 @@ open System.Reflection
 
 [<assembly: AssemblyTitleAttribute("AzureTackleAnalyzer.Tests")>]
 [<assembly: AssemblyProductAttribute("AzureTackleAnalyzer")>]
-[<assembly: AssemblyVersionAttribute("0.1.1")>]
-[<assembly: AssemblyMetadataAttribute("ReleaseDate","2020-12-14T00:00:00.0000000")>]
-[<assembly: AssemblyFileVersionAttribute("0.1.1")>]
-[<assembly: AssemblyInformationalVersionAttribute("0.1.1")>]
+[<assembly: AssemblyVersionAttribute("0.3.2")>]
+[<assembly: AssemblyMetadataAttribute("ReleaseDate","2020-12-16T00:00:00.0000000")>]
+[<assembly: AssemblyFileVersionAttribute("0.3.2")>]
+[<assembly: AssemblyInformationalVersionAttribute("0.3.2")>]
 [<assembly: AssemblyMetadataAttribute("ReleaseChannel","release")>]
-[<assembly: AssemblyMetadataAttribute("GitHash","2d4374feee7ee032ab27fbfa0edb209725b5b68c")>]
+[<assembly: AssemblyMetadataAttribute("GitHash","ba713bee009ec63e1ecbc841fb2c68bbfffdeb0b")>]
 do ()
 
 module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "AzureTackleAnalyzer.Tests"
     let [<Literal>] AssemblyProduct = "AzureTackleAnalyzer"
-    let [<Literal>] AssemblyVersion = "0.1.1"
-    let [<Literal>] AssemblyMetadata_ReleaseDate = "2020-12-14T00:00:00.0000000"
-    let [<Literal>] AssemblyFileVersion = "0.1.1"
-    let [<Literal>] AssemblyInformationalVersion = "0.1.1"
+    let [<Literal>] AssemblyVersion = "0.3.2"
+    let [<Literal>] AssemblyMetadata_ReleaseDate = "2020-12-16T00:00:00.0000000"
+    let [<Literal>] AssemblyFileVersion = "0.3.2"
+    let [<Literal>] AssemblyInformationalVersion = "0.3.2"
     let [<Literal>] AssemblyMetadata_ReleaseChannel = "release"
-    let [<Literal>] AssemblyMetadata_GitHash = "2d4374feee7ee032ab27fbfa0edb209725b5b68c"
+    let [<Literal>] AssemblyMetadata_GitHash = "ba713bee009ec63e1ecbc841fb2c68bbfffdeb0b"