+++ RELEASE_NOTES.md
@@ -1,3 +1,41 @@
+#### 0.3.2 - 2020-12-16
+* try to analyse column reading
+#### 0.3.1 - 2020-12-16
+* use json file to receive connection string
+#### 0.3.0 - 2020-12-16
+* try to fix issue with reading connection string
+#### 0.2.9 - 2020-12-15
+* try to use real connection string
+#### 0.2.8 - 2020-12-15
+* list tables
+#### 0.2.7 - 2020-12-15
+* add info message
+#### 0.2.6 - 2020-12-15
+* test directly
+#### 0.2.5 - 2020-12-15
+* get tables correctly
+#### 0.2.4 - 2020-12-15
+* optimize Error Message
+#### 0.2.3 - 2020-12-15
+* get tables from StorageAccount
+#### 0.2.2 - 2020-12-15
+* fix table
+#### 0.2.1 - 2020-12-15
+* Add Lambda
+#### 0.2.0 - 2020-12-15
+* Print Blocks
+#### 0.1.9 - 2020-12-15
+* Print Connection String
+#### 0.1.8 - 2020-12-15
+* Only analyse
+#### 0.1.7 - 2020-12-15
+* Bring back analyzer
+#### 0.1.6 - 2020-12-15
+* Migrate to one project
+#### 0.1.5 - 2020-12-15
+* Use test analyzer for now
+#### 0.1.4 - 2020-12-15
+* new release
 #### 0.1.3 - 2020-12-15
 * New push
 #### 0.1.2 - 2020-12-15